[REF] Change epochs to be indexed from 1
It doesn't make much sense for the loader to return one more than
the number of checkpointing steps taken, so now the loader returns
the number of checkpointing steps instead. This necessitates indexing
epochs from 1 to distinguish between no epochs performed (no file
loaded), which returns 0 and 1 epoch performed (returns 1).
This change is desirable anyway because thinking about epochs in
this way means it is natural for 0.5 to be halfway through the first
epoch, and 0 to be the initial model before any training.